refactor(infra): разделён AppActionRouter по action areas
AppActionRouter оставлен общей app-level точкой доступа к dependencies
navigation, source management, refresh, cleanup/storage и badge actions вынесены в отдельные extensions
